Move save_sd[xl]_lora_checkpoint(...) utility functions to a more appropriate scope.
